refactoring bridge interfaces
